support beta in float32

+++ fla/ops/gated_delta_rule/wy_fast.py
@@ -225,7 +225,7 @@ def prepare_wy_repr_bwd_kernel(
         b_A += tl.dot(b_k, b_kt)
         b_dkb = tl.dot(b_dA, b_k)
         b_db += tl.sum(b_dkb * b_k, 1)
-        b_dk = b_dkb * b_b[:, None] + tl.trans(tl.dot(b_ktb, b_dA))
+        b_dk = b_dkb * b_b[:, None] + tl.trans(tl.dot(b_ktb.to(b_dA.dtype), b_dA))
         b_dk += tl.load(p_dk, boundary_check=(0, 1))
 
         tl.store(p_dk, b_dk.to(p_dk.dtype.element_ty), boundary_check=(0, 1))
